update login component
