+++ src/app/login/page.jsx
@@ -2,6 +2,10 @@ import React from "react";
 import { Login } from "../../components/Login";
 
 const Page = () => {
-  return <Login />
-  };
+  return (
+    <div className="flex justify-center items-center bg-[#68BCE3] w-full static">
+      <Login />
+    </div>
+  );
+};
 export default Page;
